Вынес логику из функций
