+++ src/view/point-add-edit.js
@@ -2,20 +2,21 @@ import dayjs from 'dayjs';
 import { allOffers, getAvailableOffers } from '../mock/points.js';
 
 
-const showOffers = (availableOffers, selectedOffers) => ((availableOffers.length !== 0) ? `<section class="event__section  event__section--offers">
-  <h3 class="event__section-title  event__section-title--offers">Offers</h3>
-  <div class="event__available-offers">${availableOffers.map((offer) => `<div class="event__offer-selector">
-  <input class="event__offer-checkbox  visually-hidden" id="event-offer-${offer.title.split(' ').pop()}-1" type="checkbox" name="event-offer-${offer.title.split(' ').pop()}" ${(selectedOffers.find((item) => item.title === offer.title)) ? 'checked' : ''} >
-  <label class="event__offer-label" for="event-offer-${offer.title.split(' ').pop()}-1">
+const showOffers = (availableOffers, selectedOffers) =>
+  `<section class="event__section  event__section--offers">
+    <h3 class="event__section-title  event__section-title--offers">Offers</h3>
+    <div class="event__available-offers">${availableOffers.map((offer) => `<div class="event__offer-selector">
+    <input class="event__offer-checkbox  visually-hidden" id="event-offer-${offer.title.split(' ').pop()}-1" type="checkbox" name="event-offer-${offer.title.split(' ').pop()}" ${(selectedOffers.find((item) => item.title === offer.title)) ? 'checked' : ''} >
+    <label class="event__offer-label" for="event-offer-${offer.title.split(' ').pop()}-1">
     <span class="event__offer-title">${offer.title}</span>
     &plus;&euro;&nbsp;
     <span class="event__offer-price">${offer.price}</span>
-  </label>
-</div>`).join('\n')}
-  </div>
-</section>` : '');
+    </label>
+    </div>`).join('\n')}
+    </div>
+  </section>`;
 
-const showDestination = ({name = '', description = '', pictures = ''}) => (name) ? (
+const showDestination = ({description = '', pictures = ''}) =>
   `<section class="event__section  event__section--destination">
     <h3 class="event__section-title  event__section-title--destination">Destination</h3>
     <p class="event__destination-description">${description}</p>
@@ -26,8 +27,7 @@ const showDestination = ({name = '', description = '', pictures = ''}) => (name)
     `<img class="event__photo" src="${picture.src}" alt="${picture.description}">`).join('\n')}
       </div>
     </div>` : ''}
-  </section>`) : '';
-
+  </section>`;
 
 export const createAddEditPointTemplate = (point = {}, edited = false) => {
   const {basePrice = '', dateFrom = dayjs(), dateTo = dayjs(), eventType = 'Flight', eventOffers = [], destination = {}} = point;
@@ -134,9 +134,9 @@ export const createAddEditPointTemplate = (point = {}, edited = false) => {
         </button>` : ''}
       </header>
       <section class="event__details">
-        ${showOffers(getAvailableOffers(eventType, allOffers), eventOffers)}
+        ${getAvailableOffers(eventType, allOffers).length !== 0 ? showOffers(getAvailableOffers(eventType, allOffers), eventOffers) : ''}
 
-        ${showDestination(destination)}
+        ${Object.keys(destination).length !== 0 ? showDestination(destination) : ''}
       </section>
     </form>
   </li>`;
