+++ src/view/trip-info-cost.js
@@ -1,6 +1,7 @@
-const getTripCost = (sum, point) => sum + point.basePrice + point.eventOffers.reduce((acc, offer) => acc + offer.price, 0);
+const getEventCost = (sum, point) => sum + point.basePrice + point.eventOffers.reduce((acc, offer) => acc + offer.price, 0);
 
 
 export const createTripCostTemplate = (points) => `<p class="trip-info__cost">
-  Total: &euro;&nbsp;<span class="trip-info__cost-value">${(points.length !== 0) ? points.reduce(getTripCost, 0) : 0}</span>
+  Total: &euro;&nbsp;
+  <span class="trip-info__cost-value">${(points.length !== 0) ? points.reduce(getEventCost, 0) : 0}</span>
   </p>`;
